+++ Find.h
@@ -5,7 +5,7 @@
 class Find: public FU
 {
 	public:
-	void ProgFU(int MK, LoadPoint Load, FU* Sender = nullptr) override;
+	void ProgFU(long int MK, LoadPoint Load, FU* Sender = nullptr) override;
 	FU* Copy() override; // Программа копирования ФУ
 	FU* TypeCopy() override; // Создать ФУ такого же типа (не копируя контекст
 	Find(FU *BusContext, FU *Templ) : FU(BusContext)
