+++ Gauss.h
@@ -5,7 +5,7 @@
 
 class GaussPoint : public FU {
 public:
-	void ProgFU(int MK, LoadPoint Load, FU* Sender = nullptr) override;
+	void ProgFU(long int MK, LoadPoint Load, FU* Sender = nullptr) override;
 	GaussPoint(FU* BusContext, FU* Templ) : FU(BusContext) { Bus = BusContext; };
 	GaussPoint() : FU() { Bus = nullptr; };
 	double a = 0;
@@ -23,12 +23,12 @@ private:
 
 class Gauss : public FU {
 public:
-	void ProgFU(int MK, LoadPoint Load, FU* Sender = nullptr) override;
+	void ProgFU(long int MK, LoadPoint Load, FU* Sender = nullptr) override;
 //	FU* Copy() override; // Программа копирования ФУ
 //	FU* TypeCopy() override; // Создать ФУ такого же типа (не копируя контекст)
 	Gauss(FU* BusContext, FU* Templ) : FU(BusContext) { Bus = BusContext; FUtype = 16; };
 	Gauss() : FU() { Bus = nullptr; };
-	int Size = 10; // Размерность матрицы
+	long int Size = 10; // Размерность матрицы
 	double Max = 10, Min = 10;// Максимальная и минимальная значения коээфициентов и свободных членов
 private:
 		vector< vector<GaussPoint> > Field;
