+++ MatPlot.h
@@ -9,7 +9,7 @@ private:
 	vector< vector<double> > Mesh; //  Сетка для рисования
 	vector< vector<double> > Z; //  2-мерный график
 public:
-	void ProgFU(int MK, LoadPoint Load, FU* Sender = nullptr) override;
+	void ProgFU(long int MK, LoadPoint Load, FU* Sender = nullptr) override;
 	FU* Copy() override; // Программа копирования ФУ
 	FU* TypeCopy() override; // Создать ФУ такого же типа (не копируя контекст
 	MatPlot(FU* BusContext, FU* Templ)
